black + ruff + use ruff in CI

+++ examples/advanced/fragment_picking.py
@@ -24,17 +24,18 @@ class FragmentPicking(moderngl_window.WindowConfig):
     The normal is then used to hide points that point
     away from the camera.
     """
+
     title = "Fragment Picking"
     gl_version = 3, 3
     window_size = 1280, 720
     aspect_ratio = None
     resizable = True
-    resource_dir = (Path(__file__) / '../../resources').resolve()
+    resource_dir = (Path(__file__) / "../../resources").resolve()
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         print("window buffer size:", self.wnd.buffer_size)
-        self.marker_file = Path('markers.bin')
+        self.marker_file = Path("markers.bin")
         # Object rotation
         self.x_rot = 0
         self.y_rot = 0
@@ -42,7 +43,7 @@ class FragmentPicking(moderngl_window.WindowConfig):
         self.zoom = 0
 
         # Load scene cached to speed up loading!
-        self.scene = self.load_scene('scenes/fragment_picking/centered.obj', cache=True)
+        self.scene = self.load_scene("scenes/fragment_picking/centered.obj", cache=True)
         # Grab the raw mesh/vertexarray
         self.mesh = self.scene.root_nodes[0].mesh.vao
         self.mesh_texture = self.scene.root_nodes[0].mesh.material.mat_texture.texture
@@ -58,9 +59,9 @@ class FragmentPicking(moderngl_window.WindowConfig):
         # RGBA color/diffuse layer
         self.offscreen_diffuse = self.ctx.texture(self.wnd.buffer_size, 4)
         # Textures for storing normals (16 bit floats)
-        self.offscreen_normals = self.ctx.texture(self.wnd.buffer_size, 4, dtype='f2')
+        self.offscreen_normals = self.ctx.texture(self.wnd.buffer_size, 4, dtype="f2")
         # Texture for storing the view positions rendered to framebuffer
-        self.offscreen_viewpos = self.ctx.texture(self.wnd.buffer_size, 4, dtype='f4')
+        self.offscreen_viewpos = self.ctx.texture(self.wnd.buffer_size, 4, dtype="f4")
         # Texture for storing depth values
         self.offscreen_depth = self.ctx.depth_texture(self.wnd.buffer_size)
         # Create a framebuffer we can render to
@@ -77,7 +78,7 @@ class FragmentPicking(moderngl_window.WindowConfig):
         # temporary so we can use it as a normal texture
         self.depth_sampler = self.ctx.sampler(
             filter=(moderngl.LINEAR, moderngl.LINEAR),
-            compare_func='',
+            compare_func="",
         )
 
         # A fullscreen quad just for rendering offscreen textures to the window
@@ -85,37 +86,41 @@ class FragmentPicking(moderngl_window.WindowConfig):
 
         # --- Shaders
         # Simple program just rendering texture
-        self.texture_program = self.load_program('programs/fragment_picking/texture.glsl')
-        self.texture_program['texture0'].value = 0
+        self.texture_program = self.load_program("programs/fragment_picking/texture.glsl")
+        self.texture_program["texture0"].value = 0
         # Geomtry shader writing to two offscreen layers (color, normal) + depth
-        self.geometry_program = self.load_program('programs/fragment_picking/geometry.glsl')
-        self.geometry_program['texture0'].value = 0  # use texture channel 0
+        self.geometry_program = self.load_program("programs/fragment_picking/geometry.glsl")
+        self.geometry_program["texture0"].value = 0  # use texture channel 0
 
         # Shader for linearizing depth (debug visualization)
-        self.linearize_depth_program = self.load_program('programs/linearize_depth.glsl')
-        self.linearize_depth_program['texture0'].value = 0
-        self.linearize_depth_program['near'].value = self.projection.near
-        self.linearize_depth_program['far'].value = self.projection.far
+        self.linearize_depth_program = self.load_program("programs/linearize_depth.glsl")
+        self.linearize_depth_program["texture0"].value = 0
+        self.linearize_depth_program["near"].value = self.projection.near
+        self.linearize_depth_program["far"].value = self.projection.far
 
         # Shader for picking the world position of a fragment
-        self.fragment_picker_program = self.load_program('programs/fragment_picking/picker.glsl')
-        self.fragment_picker_program['position_texture'].value = 0  # Read from texture channel 0
-        self.fragment_picker_program['normal_texture'].value = 1  # Read from texture channel 1
-        self.fragment_picker_program['diffuse_texture'].value = 2  # Read from texture channel 2
+        self.fragment_picker_program = self.load_program("programs/fragment_picking/picker.glsl")
+        self.fragment_picker_program["position_texture"].value = 0  # Read from texture channel 0
+        self.fragment_picker_program["normal_texture"].value = 1  # Read from texture channel 1
+        self.fragment_picker_program["diffuse_texture"].value = 2  # Read from texture channel 2
 
         # Picker geometry
-        self.marker_byte_size = 7 * 4  # position + normal + temperature (7 x 32bit floats) 
+        self.marker_byte_size = 7 * 4  # position + normal + temperature (7 x 32bit floats)
         self.picker_output = self.ctx.buffer(reserve=self.marker_byte_size)
         self.picker_vao = VAO(mode=moderngl.POINTS)
 
         # Shader for rendering markers
-        self.marker_program = self.load_program('programs/fragment_picking/markers.glsl')
-        self.marker_program['color'].value = 1.0, 0.0, 0.0, 1.0
+        self.marker_program = self.load_program("programs/fragment_picking/markers.glsl")
+        self.marker_program["color"].value = 1.0, 0.0, 0.0, 1.0
 
         # Marker geometry
-        self.marker_buffer = self.ctx.buffer(reserve=self.marker_byte_size * 1000)  # Resever room for 1000 points
+        self.marker_buffer = self.ctx.buffer(
+            reserve=self.marker_byte_size * 1000
+        )  # Resever room for 1000 points
         self.marker_vao = VAO(name="markers", mode=moderngl.POINTS)
-        self.marker_vao.buffer(self.marker_buffer, '3f 3f 1f', ['in_position', 'in_normal', 'temperature'])
+        self.marker_vao.buffer(
+            self.marker_buffer, "3f 3f 1f", ["in_position", "in_normal", "temperature"]
+        )
         self.num_markers = 0
 
         # Debug geometry
@@ -135,8 +140,8 @@ class FragmentPicking(moderngl_window.WindowConfig):
         self.offscreen.use()
 
         # Render the scene
-        self.geometry_program['modelview'].write(self.modelview)
-        self.geometry_program['projection'].write(self.projection.matrix)
+        self.geometry_program["modelview"].write(self.modelview)
+        self.geometry_program["projection"].write(self.projection.matrix)
         self.mesh_texture.use(location=0)  # bind texture from obj file to channel 0
         self.depth_sampler.use(location=0)
         self.mesh.render(self.geometry_program)  # render mesh
@@ -153,8 +158,8 @@ class FragmentPicking(moderngl_window.WindowConfig):
         # Render markers
         if self.num_markers > 0:
             self.ctx.point_size = 6.0  # Specify fragment size of the markers
-            self.marker_program['modelview'].write(self.modelview)
-            self.marker_program['projection'].write(self.projection.matrix)
+            self.marker_program["modelview"].write(self.modelview)
+            self.marker_program["projection"].write(self.projection.matrix)
             self.marker_vao.render(self.marker_program, vertices=self.num_markers)
 
         self.render_debug()
@@ -187,27 +192,31 @@ class FragmentPicking(moderngl_window.WindowConfig):
 
         # mouse coordinates starts in upper left corner
         # pixel positions starts and lower left corner
-        pos = int(x * self.wnd.pixel_ratio), int(self.wnd.buffer_height - (y * self.wnd.pixel_ratio))
+        pos = int(x * self.wnd.pixel_ratio), int(
+            self.wnd.buffer_height - (y * self.wnd.pixel_ratio)
+        )
         print("Picking mouse position", x, y)
         print("Viewport position", pos)
 
-        self.fragment_picker_program['texel_pos'].value = pos
-        self.fragment_picker_program['modelview'].write(self.modelview)
+        self.fragment_picker_program["texel_pos"].value = pos
+        self.fragment_picker_program["modelview"].write(self.modelview)
         self.offscreen_viewpos.use(location=0)
         self.offscreen_normals.use(location=1)
         self.offscreen_diffuse.use(location=2)
         self.picker_vao.transform(self.fragment_picker_program, self.picker_output, vertices=1)
 
         # Print position
-        x, y, z, nx, ny, nz, temperature = struct.unpack('7f', self.picker_output.read())
+        x, y, z, nx, ny, nz, temperature = struct.unpack("7f", self.picker_output.read())
         if z == 0.0:
-            print('Point is not on the mesh')
+            print("Point is not on the mesh")
             return
 
         print(f"Position: {x} {y} {z}")
         print(f"Normal: {nx} {ny} {nz}")
         print(f"Temperature: {round(temperature * 255)} (byte) {temperature} (float)")
-        self.marker_buffer.write(self.picker_output.read(), offset=self.marker_byte_size * self.num_markers)
+        self.marker_buffer.write(
+            self.picker_output.read(), offset=self.marker_byte_size * self.num_markers
+        )
         self.num_markers += 1
 
     def mouse_scroll_event(self, x_offset, y_offset):
@@ -219,11 +228,11 @@ class FragmentPicking(moderngl_window.WindowConfig):
         # Key presses
         if action == keys.ACTION_PRESS:
             if key == keys.F1:
-                print('Loading marker file')
+                print("Loading marker file")
                 self.load_markers(self.marker_file)
 
             if key == keys.F2:
-                print('Saving marker file')
+                print("Saving marker file")
                 self.save_markers(self.marker_file)
 
     def load_markers(self, path: Path):
@@ -231,7 +240,7 @@ class FragmentPicking(moderngl_window.WindowConfig):
         if not self.marker_file.exists():
             return
 
-        with open(self.marker_file, mode='rb') as fd:
+        with open(self.marker_file, mode="rb") as fd:
             size = self.marker_file.stat().st_size
             self.num_markers = size // self.marker_byte_size
             self.marker_buffer.write(fd.read(), offset=0)
@@ -241,9 +250,9 @@ class FragmentPicking(moderngl_window.WindowConfig):
         if self.num_markers == 0:
             return
 
-        with open('markers.bin', mode='wb') as fd:
+        with open("markers.bin", mode="wb") as fd:
             fd.write(self.marker_buffer.read(size=self.num_markers * self.marker_byte_size))
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     moderngl_window.run_window_config(FragmentPicking)
